+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/EnchantingFeature.java
@@ -7,11 +7,9 @@ import insane96mcp.iguanatweaksexpanded.module.misc.ITEDataPacks;
 import insane96mcp.iguanatweaksexpanded.setup.ITERegistries;
 import insane96mcp.iguanatweaksexpanded.setup.IntegratedPack;
 import insane96mcp.iguanatweaksexpanded.setup.registry.SimpleBlockWithItem;
-import insane96mcp.iguanatweaksexpanded.utils.LogHelper;
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.experience.PlayerExperience;
 import insane96mcp.iguanatweaksreborn.module.experience.anvils.Anvils;
-import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.insanelib.base.JsonFeature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
@@ -107,7 +105,7 @@ public class EnchantingFeature extends JsonFeature {
     @Override
     public void loadJsonConfigs() {
         super.loadJsonConfigs();
-        for (Enchantment enchantment : ForgeRegistries.ENCHANTMENTS.getValues()) {
+        /*for (Enchantment enchantment : ForgeRegistries.ENCHANTMENTS.getValues()) {
             if (EnchantmentsFeature.isEnchantmentDisabled(enchantment))
                 continue;
             int maxLvl = enchantment.getMaxLevel();
@@ -116,7 +114,7 @@ public class EnchantingFeature extends JsonFeature {
             for (int i = 1; i <= maxLvl; i++) {
                 LogHelper.debug("%s %d: %.1f", ForgeRegistries.ENCHANTMENTS.getKey(enchantment).toString(), i, getCost(enchantment, i));
             }
-        }
+        }*/
     }
 
     @SubscribeEvent
@@ -133,7 +131,7 @@ public class EnchantingFeature extends JsonFeature {
             return;
 
         if (event.getRight().isEnchanted() || (event.getLeft().isEnchanted() && event.getRight().is(Items.ENCHANTED_BOOK)))
-            event.setOutput(ItemStack.EMPTY);
+            event.setCanceled(true);
     }
 
     public void cleansedLapis(final AnvilUpdateEvent event) {