+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/NewEnchantmentsFeature.java
@@ -67,6 +67,7 @@ public class NewEnchantmentsFeature extends Feature {
 	public static final RegistryObject<Enchantment> SWIFT_STRIKE = ITERegistries.ENCHANTMENTS.register("swift_strike", SwiftStrike::new);
 	public static final RegistryObject<Enchantment> HEALTHY = ITERegistries.ENCHANTMENTS.register("healthy", Healthy::new);
 	public static final RegistryObject<Enchantment> REACH = ITERegistries.ENCHANTMENTS.register("reach", Reach::new);
+	public static final RegistryObject<Enchantment> PADDING = ITERegistries.ENCHANTMENTS.register("padding", Padding::new);
 	public static final RegistryObject<Enchantment> VINDICATION = ITERegistries.ENCHANTMENTS.register("vindication", Vindication::new);
 	public static final RegistryObject<Enchantment> LUCKY_HOOK = ITERegistries.ENCHANTMENTS.register("lucky_hook", LuckyHook::new);
 	public static final RegistryObject<Enchantment> CURSE_OF_EXPERIENCE = ITERegistries.ENCHANTMENTS.register("curse_of_experience", CurseOfExperience::new);
@@ -101,6 +102,12 @@ public class NewEnchantmentsFeature extends Feature {
 		Vindication.tryStackDamage(event.getEntity(), event.getSource(), event.getAmount());
 	}
 
+	//Run after hoes, shovels and Knockback feature
+	@SubscribeEvent(priority = EventPriority.LOWEST)
+	public void onKnockback(LivingKnockBackEvent event) {
+		Padding.reduceKnockback(event);
+	}
+
 	@SubscribeEvent
 	public void onPickUpExperience(PlayerXpEvent.PickupXp event) {
 		CurseOfTear.tearPlayerItems(event);
@@ -109,6 +116,7 @@ public class NewEnchantmentsFeature extends Feature {
 	@SubscribeEvent
 	public void onHurt(LivingHurtEvent event) {
 		Vindication.tryApplyDamage(event);
+		Padding.shouldApply(event);
 	}
 
 	@SubscribeEvent