+++ changelog.md
@@ -9,14 +9,15 @@
 * Curse of sinking. sink faster
 * Curse of the void. Chance for drops to be voided
 * ? Curse of the undead. Your helmet will slowly break in the sun
+* Curse of ender. You may randomly teleport when hit
 * Curse of steel fall. Increases fall damage (boots only)
-* Padding. Reduces knockback
 
 ## Upcoming
 * Curses!
   * Items with curses only can now be enchanted and get a 40% boost to max cost
   * Added Curse of Experience. Consumes experience when used
   * Curse of Tear. Items break when experience is picked up
+* Added padding enchantment
 * Mining charge crafting is now cheaper
 * All items in creative tabs are now sorted!
 * Fixed beeg ore veins generating less at deepslate level